move submit.js and submit.css to another file and update references to the component
