+++ src/landing/App.js
@@ -3,7 +3,7 @@ import React from 'react';
 import './responsive-styles.css'
 import Home from './landing-home'
 import Live from './live'
-import Submit from './submit'
+import Submit from '../SubmitFilm/Submit'
 import Notfound from './notfound'
 import HomePage from './src/Homepage/Home.js'
 import {BrowserRouter as Router, Route, Switch} from 'react-router-dom'
